refactor: 패키지 구조 및 코드 포맷 수정
