+++ Android/MyApplication/app/src/main/java/com/example/myapplication/data/model/ApiResponse.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.data.model
 data class ApiResponse<T>(
     val body: T?,
     val message: String?
