+++ Android/MyApplication/app/src/main/java/com/example/myapplication/data/model/TokenLoginRequest.kt
@@ -1,0 +1,6 @@
+package com.example.myapplication.data.model
+
+data class TokenLoginRequest(
+    val token: String
+)
+
