+++ Android/MyApplication/app/src/main/java/com/example/myapplication/data/model/TokenResponse.kt
@@ -1,0 +1,6 @@
+package com.example.myapplication.data.model
+
+data class TokenResponse(
+    val accessToken: String,
+    val refreshToken: String
+)
