+++ Android/MyApplication/app/src/main/res/layout/activity_authorization.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".AuthorizationActivity">
+    tools:context=".view.AuthorizationActivity">
 
     <TextView
         android:id="@+id/authorizationDisplayText"
